[Cleanup after pull Combat]
