+++ RPG Wizardry/Assets/Scripts/PlayerInput/ControlScheme.cs
@@ -1,0 +1,8 @@
+﻿namespace nl.SWEG.RPGWizardry.PlayerInput
+{
+    public enum ControlScheme
+    {
+        Controller = 1,
+        Keyboard = 0
+    }
+}
