+++ RPG Wizardry/Assets/Scripts/PlayerInput/ControlScheme.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c957679a44981c64b8d2c9d1d761c667
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
